put same things in common files
