+++ server/constants.php
@@ -1,0 +1,7 @@
+<?php
+define('OFFSET_X', 0);
+define('OFFSET_Y', 0);
+define('SCALE_Z',  3);
+define('TILE_SIZE', 256);
+define('MAX_ZOOM',  18);
+
